Include WellKnownDiagnosticTags.Unnecessary in the SA1121 diagnostic
Normally this flag is not needed because the Simplify Type Name diagnostic
provided by Visual Studio includes the tag and marks the same span. However,
The Simplify Type Name diagnostic is not reported for references to a using
alias directive, so adding the flag to SA1121 ensures the UI is correct for
all cases where this diagnostic might be reported.