Improve version stamps of builds
With this change, it is no longer necessary to make commits that change version data just to prepare for an official release (like alpha006) and revert it back immediately after so that local builds are versioned as 1.0.0-dev.

With the old scheme, since everyone's build was always 1.0.0-dev it made it impossible to consume the NuPkg produced by a build anywhere because 1.0.0-dev is a constant that doesn't actually point at any specific version. Now, anyone's non-official build will be unique with the commit ID where it was built, such as:

    1.0.0-alpha-g9a7eb6c819

In this way, anyone with a fork of the repo can build their own package, upload it to their private nuget store (myget.org, a network share, etc.) and consume it immediately.

When shipping officlal releases, simply build with the /p:UseNonZeroBuildNumber=true switch to produce a build number like this:

    1.0.15136-alpha

Note that this no longer uses the "-alpha007" syntax, with its fixed number of leading zeros. Instead, it just uses the PATCH component of the version as the incrementing integer for proper semver-sorting and easier reading that this is an alpha.
This is a safe upgrade from what we were doing because this new format is seen as 'newer' than the old one, so upgraders will get the new versions as expected.