Setup NuGet packaging for MetaCompilation
Setup NuGet packaging for the MetaCompilation project. MetaCompilation
won't actually be distributed as a NuGet package, but it's the only
project we have at the moment so it will have to serve as our test case.

This commit creates a new project, MetaCompilation.NuGet.proj, that
contains the targets and data necessary to create the three "flavors" of
NuGet packages we care about: per-build prerelease, prerelease, and
release. This does get a bit ugly as we end up with three nearly-
identical targets, one for each flavor. As yet I haven't found a good
way eliminate the duplication.

A future change will extract these targets and properties into a new
.targets file that can be reused by other packaging projects.

The Diagnostic.nuspec file has been updated to use numerous tokens, and
to explicitly pick up the the analyzer assembly from the binaries
directory.

Finally, Packaging.proj has been created to build all the packaging
projects. Currently, of course, it only builds this one.