Use SyntaxGenerator instead of SyntaxFactory
Update `CodeFixNodeCreator.NewIdCreator(...)` to use a `SyntaxGenerator`
instead of `SyntaxFactory`. The output of the current version formats in
a different way when attempting to move from the Roslyn 1.0.0-rc2
packages to the Roslyn 1.0.0-rc3 packages, causing unit tests to fail.
By using `SyntaxGenerator` in this scenario I can avoid the need to
create trivia, and leave it to the `SyntaxGenerator` to do the right
thing.