Preserve trailing trivia
The `InternalStaticAsync` method updates an existing field declaration
to have specific accessibility and modifiers. In doing so, it replaces
the trailing trivia with a single space. Under the code formatter in
Roslyn 1.0.0-rc2 this happens to produce the desired output, but it does
not do so using 1.0.0-rc3.

The correct thing to do is to simply preserve the existing trailing
trivia. That should produce the desired output regardless of changes in
the formatter.