Fix up trivia handling
This change simplifies `InternalStaticAsync(...)` to favor
`SyntaxFactory.GetToken(...)` over `SyntaxFactory.ParseToken(...)`, and
to avoid creating trivia for the new tokens. Instead, we let the
`SyntaxFactory` insert elastic trivia and let the formatting engine take
care of the resulting whitespace.

This is necessary to get all the unit tests working again.