Parallelize unit test runs
By default the xUnit console runner will run test assemblies one at a
time, but within a test assembly run test collections in parallel. Here
we update BuildAndTest.proj to both run the assemblies in parallel, and
the collections within the assemblies. This shaves a couple of seconds
off of the run time.