Add CustomBatchFixAllProvider
This code is a copy of BatchFixAllProvider that has been modified to work
in a separate assembly by replacing uses of internal APIs with public APIs.
In addition, the generated code detection mechanism was completely removed;
this detection is already performed by analyzers prior to reporting
diagnostics so duplicating the check is a waste of resources.