Make all diagnostic ids constants
The status page has to figure out what descriptor matches which Descriptor object it has in memory. It tries to get the value of the id by using SemanticModels GetConstantValue method.
To make this work the variables used have to be constant. Without this change the status page shows "Unknown" for the status of the diagnostic