Use a builder instead of ImmutableDictionary in JSON models
This works around JamesNK/Newtonsoft.Json#652 by avoiding the direct use
of immutable types in the serialization process. The result remains
efficient because the builder class caches the last returned immutable
instance, and will continue returning the same instance as long as no
changes are made to the builder.

Fixes #1359