Use legally-correct copyright notices.
The existing copyright support uses the (legally-unrecognized) (C), and does not support year lists due to the literal-matching nature of copyrightText.

This change fixes the above problems by adding a new {copyright} variable for use with copyrightText. The default value for copyrightText is also updated to use {copyright} instead of the hard-coded "Copyright (C)" text.

{copyright} provides the following features:
- When suggesting a fix, automatically expands to "Copyright © " followed by the current year.
- When checking copyright headers, {copyright} matches "Copyright © " followed by a comma-separated list of four-digit years, and/or year-ranges (two four-digit years separated with a hyphen).

This change adds new unit tests to verify this new functionality, and updates the existing tests to work correctly with the new behaviors.