Implement UseCultureAttribute
This is necessary to make sure all tests use en-US to get standard texts
without explicitly specifying the culture every time the resource file is
accessed.

The code for this class was taken from the xunit/sample.xunit repository
which is (c) 2014 Outercurve Foundation and licensed under the Apache License,
Version 2.0.