Fix handling of both a header and a separated comment
When a comment preceding a set of using directives is separated from the first
using directive by a blank line, that comment is associated with the list of
using directives as opposed to specifically tied to the first one. This means
after reordering the separated comment will continue to be separated and placed
before the first using directive of the sorted result.