Make sure culture info is persisted across methods executed from classes with UseCulture attribute
When we change the current thread's culture, that change is lost
when an async method is called. So Microsoft guys have introduced
the CultureInfo.DefaultThreadCurrentCulture property in .NET 4.5.

See https://stackoverflow.com/a/30664385/1396155

According to this SO answer, we do not have to use this property
in .NET 4.6 because async methods seem to get the same culture info
the caller method has. So if someday the test project's target
framework is updated, feel free to revert this commit.