Fixed Code Analysis issues from internal fields & functions needing to be private / documented.
These incorrect analysis issues can break builds of stable API's and not all API's should document methods labeled as "internal" (known to break the build of zlib.net for example and I do not like dirtying the code with suppressions making it hard to maintain). As such these analyzers need this fix instead.