Keep tracked nodes in a list
Lazily-enumerated nodes which are passed to TrackNodes and then
enumerated a second time can be garbage collected between the two
passes. If this occurs, the node tracking will fail. We resolve the
issue by ensuring the nodes are held in an eagerly-evaluated list.

See https://github.com/dotnet/roslyn/pull/61958